extract img manipulation into new file
